+++ selenium-debug.log
@@ -1,0 +1,54 @@
+20:08:00.072 INFO - Launching a standalone Selenium Server
+20:08:00.165 INFO - Java: Oracle Corporation 25.73-b02
+20:08:00.165 INFO - OS: Windows 10 10.0 x86
+20:08:00.179 INFO - v2.52.0, with Core v2.52.0. Built from revision 4c2593c
+20:08:00.241 INFO - Driver class not found: com.opera.core.systems.OperaDriver
+20:08:00.241 INFO - Driver provider com.opera.core.systems.OperaDriver is not registered
+20:08:00.250 INFO - Driver provider org.openqa.selenium.safari.SafariDriver registration is skipped:
+registration capabilities Capabilities [{browserName=safari, version=, platform=MAC}] does not match the current platform WIN10
+20:08:00.389 INFO - RemoteWebDriver instances should connect to: http://127.0.0.1:4444/wd/hub
+20:08:00.389 INFO - Selenium Server is up and running
+20:08:00.961 INFO - Executing: [new session: Capabilities [{acceptSslCerts=true, name=Example, browserName=firefox, javascriptEnabled=true, platform=ANY}]])
+20:08:00.972 INFO - Creating a new session for Capabilities [{acceptSslCerts=true, name=Example, browserName=firefox, javascriptEnabled=true, platform=ANY}]
+20:08:09.667 INFO - Done: [new session: Capabilities [{acceptSslCerts=true, name=Example, browserName=firefox, javascriptEnabled=true, platform=ANY}]]
+20:08:09.680 INFO - Executing: [get: http://www.google.com])
+20:08:12.255 INFO - Done: [get: http://www.google.com]
+20:08:12.263 INFO - Executing: [find elements: By.cssSelector: body])
+20:08:12.533 INFO - Done: [find elements: By.cssSelector: body]
+20:08:12.541 INFO - Executing: [is displayed: 0 [[FirefoxDriver: firefox on WINDOWS (e5a6e6de-ee5b-424e-835e-efd533616740)] -> css selector: body]])
+20:08:12.556 INFO - Done: [is displayed: 0 [[FirefoxDriver: firefox on WINDOWS (e5a6e6de-ee5b-424e-835e-efd533616740)] -> css selector: body]]
+20:08:12.563 INFO - Executing: [find element: By.cssSelector: input[type=text]])
+20:08:12.579 INFO - Done: [find element: By.cssSelector: input[type=text]]
+20:08:12.583 INFO - Executing: [send keys: 1 [[FirefoxDriver: firefox on WINDOWS (e5a6e6de-ee5b-424e-835e-efd533616740)] -> css selector: input[type=text]], [n, i, g, h, t, w, a, t, c, h]])
+20:08:12.685 INFO - Done: [send keys: 1 [[FirefoxDriver: firefox on WINDOWS (e5a6e6de-ee5b-424e-835e-efd533616740)] -> css selector: input[type=text]], [n, i, g, h, t, w, a, t, c, h]]
+20:08:12.692 INFO - Executing: [find elements: By.cssSelector: button[name=btnG]])
+20:08:12.707 INFO - Done: [find elements: By.cssSelector: button[name=btnG]]
+20:08:12.710 INFO - Executing: [is displayed: 2 [[FirefoxDriver: firefox on WINDOWS (e5a6e6de-ee5b-424e-835e-efd533616740)] -> css selector: button[name=btnG]]])
+20:08:12.729 INFO - Done: [is displayed: 2 [[FirefoxDriver: firefox on WINDOWS (e5a6e6de-ee5b-424e-835e-efd533616740)] -> css selector: button[name=btnG]]]
+20:08:12.734 INFO - Executing: [find element: By.cssSelector: button[name=btnG]])
+20:08:12.769 INFO - Done: [find element: By.cssSelector: button[name=btnG]]
+20:08:12.773 INFO - Executing: [click: 2 [[FirefoxDriver: firefox on WINDOWS (e5a6e6de-ee5b-424e-835e-efd533616740)] -> css selector: button[name=btnG]]])
+20:08:12.890 INFO - Done: [click: 2 [[FirefoxDriver: firefox on WINDOWS (e5a6e6de-ee5b-424e-835e-efd533616740)] -> css selector: button[name=btnG]]]
+20:08:13.897 INFO - Executing: [find element: By.cssSelector: #main])
+20:08:14.341 INFO - Done: [find element: By.cssSelector: #main]
+20:08:14.344 INFO - Executing: [get text: 3 [[FirefoxDriver: firefox on WINDOWS (e5a6e6de-ee5b-424e-835e-efd533616740)] -> css selector: #main]])
+20:08:15.796 INFO - Done: [get text: 3 [[FirefoxDriver: firefox on WINDOWS (e5a6e6de-ee5b-424e-835e-efd533616740)] -> css selector: #main]]
+20:08:15.803 INFO - Executing: [delete session: 5074e941-ed1e-46bb-aed0-89e958b4d687])
+20:08:16.532 INFO - Done: [delete session: 5074e941-ed1e-46bb-aed0-89e958b4d687]
+20:08:16.655 INFO - Executing: [new session: Capabilities [{acceptSslCerts=true, name=Front Page, browserName=firefox, javascriptEnabled=true, platform=ANY}]])
+20:08:16.655 INFO - Creating a new session for Capabilities [{acceptSslCerts=true, name=Front Page, browserName=firefox, javascriptEnabled=true, platform=ANY}]
+20:08:22.760 INFO - Done: [new session: Capabilities [{acceptSslCerts=true, name=Front Page, browserName=firefox, javascriptEnabled=true, platform=ANY}]]
+20:08:22.764 INFO - Executing: [get: file:C:\Users\nkula\Desktop\workspace\eda-sim-front\dist\index.html])
+20:08:25.573 INFO - Done: [get: file:C:\Users\nkula\Desktop\workspace\eda-sim-front\dist\index.html]
+20:08:25.579 INFO - Executing: [find elements: By.cssSelector: body])
+20:08:25.644 INFO - Done: [find elements: By.cssSelector: body]
+20:08:25.649 INFO - Executing: [is displayed: 0 [[FirefoxDriver: firefox on WINDOWS (efc6f873-636b-430e-9739-d9e70ffc3c8f)] -> css selector: body]])
+20:08:25.663 INFO - Done: [is displayed: 0 [[FirefoxDriver: firefox on WINDOWS (efc6f873-636b-430e-9739-d9e70ffc3c8f)] -> css selector: body]]
+20:08:26.667 INFO - Executing: [find elements: By.cssSelector: .phase])
+20:08:26.682 INFO - Done: [find elements: By.cssSelector: .phase]
+20:08:26.686 INFO - Executing: [find elements: By.cssSelector: div.status])
+20:08:26.702 INFO - Done: [find elements: By.cssSelector: div.status]
+20:08:26.706 INFO - Executing: [get text: 2 [[FirefoxDriver: firefox on WINDOWS (efc6f873-636b-430e-9739-d9e70ffc3c8f)] -> css selector: div.status]])
+20:08:26.732 INFO - Done: [get text: 2 [[FirefoxDriver: firefox on WINDOWS (efc6f873-636b-430e-9739-d9e70ffc3c8f)] -> css selector: div.status]]
+20:08:26.741 INFO - Executing: [delete session: b1aeaa1d-71db-4f0b-9081-af0ef56e6714])
+20:08:27.044 INFO - Done: [delete session: b1aeaa1d-71db-4f0b-9081-af0ef56e6714]
