+++ src/data/initial-state.js
@@ -3,7 +3,7 @@ let initialState = Map({
   isPlaying: false,
   score: 2435,
   money: 200,
-  gameover: false,
+  gameover: true,
   activeTask: '',
   time: Map({
     hour: 7,
